chore: fixes Sentry warning to happen in dev mode + move Sentry code to .ts
